+++ ChoGGi's Library/metadata.lua
@@ -1,7 +1,7 @@
 return PlaceObj("ModDef", {
-	"title", "ChoGGi's Library v4.9 Test",
-	"version", 49,
-	"saved", 1544875200,
+	"title", "ChoGGi's Library v5.0 Test",
+	"version", 50,
+	"saved", 1545566400,
 	"image", "Preview.png",
 	"id", "ChoGGi_Library",
 	"steam_id", "1504386374",